fix: stop pinning vue peer dependency version
It should be safe to trust `vue` to follow semver.
Pinning a peer dependency version is a bad idea because any change in
peer dependency requirement should be considered a breaking change in
theory.

On the other hand, `@vue/compiler-sfc` has a strict peer dependency
version requirement.
So pinning the `vue` version is likely to causing issues like
https://github.com/vitejs/vite/issues/4573 in the future.