fix(components): [dropdown] fix #8311
* dropdown/src/content.vue: `triggeringElementRef` is a reference of
`el-button` component, not the `button` dom object. We should get its
dom object before passing it to `virtual-ref`.
* popper/src/content.vue: `referenceEl` is `triggeringElementRef`, which
is a dom object. Because both dom object and its instance function
`getBoundingClientRect` are not reactive, watching their changes are of
non-sence.

It's strange that it worked before v2.2.6, don't know what happened