fix(components): [table] fixed an undefined reference in table-layout
The table layout's `updateScrollY` function could throw an error by referencing a property on an undefined object. The scroll bar ref exposes `wrapRef` which could be undefined at the time that `scrollHeight` is accessed to determine a new `scrollY` value. This could be reproduced in tests where the `updateScrollY` function is called asynchronously, but executes after the test's component wrapper has already been unmounted.