fix(components): [select] ensure select options respect filter query during initialization (#11205)
* fix(components): [select] ensure select options respect query immediately during initialization

When the select option is created, the option does not immediately calculate its `visible` state attribute. This can result in the visible select options being out of sync with the filter text. If the options are loaded asynchronously and a user is quick to enter filter text, then the options will show regardless of the filter value once they have completed loading. Fix this by updating the `visible` attribute immediately.

* fix: lint

---------

Co-authored-by: zz <2418184580@qq.com>