feat(components): [tag] Added custom tag color functionality
Added the ability to customize the color of the tag component
by introducing the customTagColor prop.
If the props.color is set, it will be used as the background color.
If not, the props.customTagColor will be used to calculate the background, border,
and text colors for three different effects: dark, light, and plain.
This enhancement allows for more flexibility in styling the tag component,
enabling users to create visually appealing and customizable tags.

- Implemented custom tag color functionality
- Added support for customTagColor prop
- Improved styling options for tag component