fix(components): [el-slider] fix v-model value is out of range
When the default v-model value is out of the range defined by the minimum and maximum values, the UI
will only display the nearest valid maximum or minimum value.

However, the v-model itself is not updated, which leads to inconsistencies between the UI and the
data,

potentially causing confusion.

BREAKING CHANGE :
There are no breaking changes introduced by this fix.