fix(components): [el-cascader] fix loading when setting lazy and v-model
When using lazyLoad,

after setting the default value with v-model to load the child nodes of the corresponding node,

even after loading is complete, the loading icon does not disappear.

The loading icon only disappears when the node is clicked to expand.

BREAKING CHANGE :
There are no breaking changes introduced by this fix

closed Fixes #17655