docs: fix typo and improve sentence in destroy-on-close example (#22630)
docs(docs): fix typo and improve sentence in destroy-on-close example

When I was reading the guide documentation,
I found this sentence a bit confusing.
It turned out that there was a small typo (“bellow” → “below”)
and the phrasing wasn’t very natural.
So I tried to fix it for better readability.
Thanks to all the maintainers for your hard work!

Co-authored-by: Linzx <14808625+sheldorplus@user.noreply.gitee.com>